+++ saitynai-server/Controllers/CommentsController.cs
@@ -1,7 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using saitynai_server.Auth.Model;
 using saitynai_server.Data.Dtos.Comments;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 
 namespace saitynai_server.Controllers
 {
@@ -10,19 +14,22 @@ namespace saitynai_server.Controllers
     public class CommentsController : ControllerBase
     {
         private readonly IMapper _mapper;
+        private readonly IAuthorizationService _authorizationService;
         private readonly IGamesRepository _gamesRepository;
         private readonly IAdvertisementsRepository _advertisementsRepository;
         private readonly ICommentsRepository _commentsRepository;
 
-        public CommentsController(IMapper mapper, IGamesRepository gamesRepository, IAdvertisementsRepository advertisementsRepository, ICommentsRepository commentsRepository)
+        public CommentsController(IMapper mapper, IAuthorizationService authorizationService, IGamesRepository gamesRepository, IAdvertisementsRepository advertisementsRepository, ICommentsRepository commentsRepository)
         {
             _mapper = mapper;
+            _authorizationService = authorizationService;
             _gamesRepository = gamesRepository;
             _advertisementsRepository = advertisementsRepository;
             _commentsRepository = commentsRepository;
         }
 
         [HttpGet]
+        [AllowAnonymous]
         public async Task<ActionResult<IEnumerable<CommentDto>>> GetAll(int gameId, int advertisementId)
         {
             var game = await _gamesRepository.GetAsync(gameId);
@@ -39,6 +46,7 @@ namespace saitynai_server.Controllers
         }
 
         [HttpGet("{commentId}")]
+        [AuthorizeByRoles(Roles.Admin, Roles.User)]
         public async Task<ActionResult<CommentDto>> Get(int gameId, int advertisementId, int commentId)
         {
             var game = await _gamesRepository.GetAsync(gameId);
@@ -53,10 +61,15 @@ namespace saitynai_server.Controllers
             if (comment == null)
                 return NotFound($"Comment with fkAdvertisementId '{advertisementId}' and id '{commentId}' not found.");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, comment, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+                return Forbid(); // could be 404 for security
+
             return Ok(_mapper.Map<CommentDto>(comment));
         }
 
         [HttpPost]
+        [AuthorizeByRoles(Roles.User)]
         public async Task<ActionResult<CommentDto>> Create(int gameId, int advertisementId, CommentPostDto commentPostDto)
         {
             var game = await _gamesRepository.GetAsync(gameId);
@@ -69,7 +82,7 @@ namespace saitynai_server.Controllers
 
             var comment = _mapper.Map<Comment>(commentPostDto);
             comment.FkAdvertisement = advertisement;
-            // ==============| SET USER ID HERE |===============
+            comment.UserId = User.FindFirstValue(JwtRegisteredClaimNames.Sub);
 
             await _commentsRepository.CreateAsync(comment);
 
@@ -77,6 +90,7 @@ namespace saitynai_server.Controllers
         }
 
         [HttpPut("{commentId}")]
+        [AuthorizeByRoles(Roles.User)]
         public async Task<ActionResult<CommentDto>> Update(int gameId, int advertisementId, int commentId, CommentUpdateDto commentUpdateDto)
         {
             var game = await _gamesRepository.GetAsync(gameId);
@@ -91,6 +105,10 @@ namespace saitynai_server.Controllers
             if (oldComment == null)
                 return NotFound($"Comment with fkAdvertisementId '{advertisementId}' and id '{commentId}' not found.");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, oldComment, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+                return Forbid(); // could be 404 for security
+
             _mapper.Map(commentUpdateDto, oldComment);
 
             await _commentsRepository.UpdateAsync(oldComment);
@@ -99,6 +117,7 @@ namespace saitynai_server.Controllers
         }
 
         [HttpDelete("{commentId}")]
+        [AuthorizeByRoles(Roles.Admin, Roles.User)]
         public async Task<ActionResult> Remove(int gameId, int advertisementId, int commentId)
         {
             var game = await _gamesRepository.GetAsync(gameId);
@@ -112,6 +131,10 @@ namespace saitynai_server.Controllers
             var comment = await _commentsRepository.GetAsync(advertisementId, commentId);
             if (comment == null)
                 return NotFound($"Comment with fkAdvertisementId '{advertisementId}' and id '{commentId}' not found.");
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, comment, PolicyNames.ResourceOwner);
+            if (!authorizationResult.Succeeded)
+                return Forbid(); // could be 404 for security
 
             await _commentsRepository.DeleteAsync(comment);
